P1_Fish_And_Fly/requirements.txt
torch
torchvision
transformers
ultralytics
opencv-python
pybullet
numpy
matplotlib
pyyaml
pathlib
python-box
ensure